grub: new package at 2.04
The only bootloader that is actively developed (outside of systemd-boot and
rEFInd, but the former may require glibc and the latter depends on bash in
runtime, so Wyvertux is out of question to use the latter). So, provide the
option. While Wyvertux aims to remove GNU when possible, we have to be
pragmatic.

(I am using GRUB myself)